erstydfguh

+++ Libbrerie/SensoreSeguiLinea/src/SensoreSeguiLinea.h
@@ -6,7 +6,7 @@ class SensoreSeguiLinea {
     public:
         SensoreSeguiLinea(uint8_t adress = 0x20);
         void begin();
-        int read(uint8_t reg);
+        uint8_t read(uint16_t reg);
     private:
         uint8_t _adress;
 };
